% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amazonia_33K_12classes_4bands}
\alias{amazonia_33K_12classes_4bands}
\title{Samples of Amazonia biome}
\format{A \code{sits} tibble with 33K samples. A \code{sits} tibble contains the following meta-data:
  longitude (East-west coordinate of the time series sample in WGS 84),
  latitude (North-south coordinate of the time series sample in WGS 84),
  start_date (initial date of the time series),
  end_date (final date of the time series),
  label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data), and
  time_series ( list containing a tibble with the values of the time series).

  The samples are from the work of Ana Rorato, combined with agricultural data provided by EMBRAPA.
  There are samples of 12 classes ("Fallow_Cotton", "Forest", "Millet_Cotton", "Pasture", "Savanna",
  "Savanna_Roraima", "Soy_Corn", "Soy_Cotton", "Soy_Fallow", "Soy_Millet", "Soy_Sunflower" and "Wetlands").

  Each time series covers 12 months (23 data points) from the MOD13Q1 product,
  in 4 bands ("ndvi", "evi", "nir", "mir").}
\usage{
data("amazonia_33K_12classes_4bands")
}
\description{
A dataset containing a sits tibble with time series samples from Brazilian Amazonia biome.
}
