% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cerrado_124K_16classes_6bands}
\alias{cerrado_124K_16classes_6bands}
\title{Samples of Cerrado biome}
\format{A `sits` tibble with 124K samples. A `sits` tibble contains the following meta-data: 
  longitude (East-west coordinate of the time series sample in WGS 84),
  latitude (North-south coordinate of the time series sample in WGS 84), 
  start_date (initial date of the time series),
  end_date (final date of the time series), 
  label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data), and
  time_series ( list containing a tibble with the values of the time series).
  
  The samples are from many sources. It has 16 classes (_Araguaia_, _Campo\_Cerrado_, 
  _Cerradao_, _Cerrado_, _Cerrado\_Rupestre_, _Ciliary\_Forest_, _Dunas_, 
  _Fallow\_Cotton_, _Millet\_Cotton_, _Pasture_, _Perennial\_Crop_, 
  _Semi\_Perennial\_Crop_, _Soy\_Corn_, _Soy\_Cotton_, _Soy\_Fallow_, _Soy\_Millet_).
  
  Each time series comprehends 12 months (23 data points) from MOD13Q1 product, and has 6 bands.}
\usage{
data("cerrado_124K_16classes_6bands")
}
\description{
A dataset containing a sits tibble with time series samples from Brazilian Cerrado biome.
}
