% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_datasets}
\alias{list_datasets}
\title{List data sets}
\usage{
list_datasets()
}
\value{
character vector with all data sets names.
}
\description{
Lists all data sets provided by the package.
}
\examples{
list_datasets()

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
