% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_style.R
\name{qgis_style}
\alias{qgis_style}
\title{Predefined QGIS Style}
\usage{
qgis_style(dataset)
}
\arguments{
\item{dataset}{(string) one valid data set provided by inSitu.}
}
\value{
nothing.
}
\description{
Print a predefined QGIS Style (.qml) markup text on screen.
Copy and paste to create a .qml file to be used in QGIS.
The style is associated with a given data set.
}
\examples{
qgis_style("cerrado_124K_16classes_6bands")

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
